% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_width_ci.R
\name{get_width_ci}
\alias{get_width_ci}
\title{Title Confidence intervals width with estimates}
\usage{
get_width_ci(est, iso_code, coverage = 95)
}
\arguments{
\item{est:}{tibble which contains mCPR estimates. Columns: “Country or area”, iso, Year, Median, U95,L95}

\item{iso_code:}{country iso code}

\item{coverage:}{width of intervals to be calculated. Options can be: 80, 95, or NA}
}
\value{
: a tibble with Year and interval widths
}
\description{
Title Confidence intervals width with estimates
}
\examples{
: function(est,iso_code , coverage = 95)
}
