% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cp.R
\name{plot_cp}
\alias{plot_cp}
\title{Title ggplot with data and estimates or error messages}
\usage{
plot_cp(dat, est, iso_code, CI = 95)
}
\arguments{
\item{dat:}{tibble which contains mCPR observations. Columns: iso, year, cp}

\item{est:}{tibble which contains mCPR estimates. Columns: “Country or area”, iso, Year, Median, U95,L95}

\item{iso_code:}{country iso code}

\item{CI:}{confidence intervals to be plotted. Options can be: 80, 95, or NA (no CI plotted)}
}
\value{
: Either error messages or ggplot object with data and estimates
}
\description{
Title ggplot with data and estimates or error messages
}
\examples{
: function(dat, est, iso_code, CI=95)
}
